requirements.txt
torch
tensorflow
numpy
psutil
pyyaml
pytest
